app/src/server/requirements.txt
MarkupSafe==3.0.2
Pillow==11.3.0
PyYAML==6.0.2
SQLAlchemy==2.0.42
accelerate>=0.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2024.12.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json-repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langchain-experimental==0.3.4
langchain-neo4j==0.5.0
langchain-ollama==0.3.6
langsmith==0.4.13
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy-extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.9.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
ollama==0.5.2
openai==1.99.1
orjson==3.11.1
packaging==25.0
propcache==0.3.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pypdf==5.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.6.1
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.10.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
triton==3.4.0
types-PyYAML==6.0.12.20250516
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
yandex_cloud_ml_sdk==0.13.1
zstandard==0.23.0